+++ e2e/uni-builder/inline-chunk/src/another.js
@@ -1,9 +1,0 @@
-import './style.css';
-
-import(
-  /* webpackChunkName: "foo" */
-  './foo'
-).then(({ foo }) => {
-  // eslint-disable-next-line no-undef
-  window.answer = `another ${foo()}`;
-});