+++ e2e/uni-builder/resolve-extension-prefix/src/ex.js
@@ -1,5 +1,0 @@
-/* eslint-disable no-undef */
-const testEl = document.createElement('div');
-testEl.id = 'test-el';
-testEl.innerHTML = 'aaaaa';
-document.body.appendChild(testEl);