+++ e2e/uni-builder/sri/src/App.jsx
@@ -1,9 +1,0 @@
-import './App.css';
-
-const App = () => (
-  <div className="container">
-    <div id="test">Hello Rsbuild!</div>
-  </div>
-);
-
-export default App;