+++ packages/core/src/provider/plugins/minimize.ts
@@ -1,38 +1,5 @@
-import { CHAIN_ID, type RspackBuiltinsConfig } from '@rsbuild/shared';
-import type { RsbuildPlugin, NormalizedConfig } from '../../types';
-
-const getJsMinimizerOptions = (config: NormalizedConfig) => {
-  const options: RspackBuiltinsConfig['minifyOptions'] = {};
-
-  const { removeConsole } = config.performance;
-
-  if (removeConsole === true) {
-    options.dropConsole = true;
-  } else if (Array.isArray(removeConsole)) {
-    const pureFuncs = removeConsole.map((method) => `console.${method}`);
-    options.pureFuncs = pureFuncs;
-  }
-
-  switch (config.output.legalComments) {
-    case 'inline':
-      options.comments = 'some';
-      options.extractComments = false;
-      break;
-    case 'linked':
-      options.extractComments = true;
-      break;
-    case 'none':
-      options.comments = false;
-      options.extractComments = false;
-      break;
-    default:
-      break;
-  }
-
-  options.asciiOnly = config.output.charset === 'ascii';
-
-  return options;
-};
+import { CHAIN_ID, getSwcMinimizerOptions } from '@rsbuild/shared';
+import type { RsbuildPlugin } from '../../types';
 
 export const pluginMinimize = (): RsbuildPlugin => ({
   name: 'rsbuild:minimize',
@@ -54,7 +21,7 @@ export const pluginMinimize = (): RsbuildPlugin => ({
 
       chain.optimization
         .minimizer(CHAIN_ID.MINIMIZER.JS)
-        .use(SwcJsMinimizerRspackPlugin, [getJsMinimizerOptions(config)])
+        .use(SwcJsMinimizerRspackPlugin, [getSwcMinimizerOptions(config)])
         .end()
         .minimizer(CHAIN_ID.MINIMIZER.CSS)
         .use(SwcCssMinimizerRspackPlugin, [])